task_enroll_user_in_pending_courses moved to /philu_overrides/tasks.py
